requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0

fastapi>=0.115.0
uvicorn[standard]>=0.30.0

langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0

faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
rank-bm25>=0.2.2
pytest>=8.0.0
